Blogflix/api/requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.4.1
billiard==3.6.3.0
boto3==1.13.12
botocore==1.16.12
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-storages==1.9.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.15.2
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.10.0
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
Pillow==7.1.1
pylint==2.5.2
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.0
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
